requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
cryptography==46.0.5
ecdsa==0.19.2
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMySQL==1.1.2
pytest==9.0.2
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyJWT==2.9.0 
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
